adding validation, role_admin, fixes

+++ src/Controller/CourseController.php
@@ -12,6 +12,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+
 
 /**
  * @Route("/course")
@@ -32,8 +35,12 @@ class CourseController extends AbstractController
     /**
      * @Route("/new", name="course_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(Request $request, AuthorizationCheckerInterface $authChecker): Response
     {
+        if (false === $authChecker->isGranted('ROLE_SUPER_ADMIN')) {
+            return $this->redirectToRoute('course_index');
+        }
+
         $course = new Course();
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
@@ -71,8 +78,12 @@ class CourseController extends AbstractController
     /**
      * @Route("/{id}/edit", name="course_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Course $course): Response
+    public function edit(Request $request, Course $course, AuthorizationCheckerInterface $authChecker): Response
     {
+        if (false === $authChecker->isGranted('ROLE_SUPER_ADMIN')) {
+            return $this->redirectToRoute('course_index');
+        }
+
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -91,8 +102,12 @@ class CourseController extends AbstractController
     /**
      * @Route("/{id}", name="course_delete", methods={"POST"})
      */
-    public function delete(Request $request, Course $course): Response
+    public function delete(Request $request, Course $course, AuthorizationCheckerInterface $authChecker): Response
     {
+        if (false === $authChecker->isGranted('ROLE_SUPER_ADMIN')) {
+            return $this->redirectToRoute('course_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($course);
